Change drawables dimensions and positions from int to float

+++ src/main/java/openperipheral/addons/glasses/IDrawableContainer.java
@@ -31,28 +31,28 @@ public interface IDrawableContainer {
 
 	@ScriptCallable(returnTypes = ReturnType.OBJECT, description = "Add a new text object to the screen")
 	public Drawable addText(
-			@Arg(name = "x", description = "The x position from the top left") short x,
-			@Arg(name = "y", description = "The y position from the top left") short y,
+			@Arg(name = "x", description = "The x position from the top left") float x,
+			@Arg(name = "y", description = "The y position from the top left") float y,
 			@Arg(name = "text", description = "The text to display") String text,
 			@Optionals @Arg(name = "color", description = "The text color") Integer color
 			);
 
 	@ScriptCallable(returnTypes = ReturnType.OBJECT, description = "Add a new box to the screen")
 	public Drawable addBox(
-			@Arg(name = "x", description = "The x position from the top left") short x,
-			@Arg(name = "y", description = "The y position from the top left") short y,
-			@Arg(name = "width", description = "The width of the box") short width,
-			@Arg(name = "height", description = "The height of the box") short height,
+			@Arg(name = "x", description = "The x position from the top left") float x,
+			@Arg(name = "y", description = "The y position from the top left") float y,
+			@Arg(name = "width", description = "The width of the box") float width,
+			@Arg(name = "height", description = "The height of the box") float height,
 			@Optionals @Arg(name = "color", description = "The color of the box") Integer color,
 			@Arg(name = "opacity", description = "The opacity of the box (from 0 to 1)") Float opacity
 			);
 
 	@ScriptCallable(returnTypes = ReturnType.OBJECT, description = "Add a new gradient box to the screen")
 	public Drawable addGradientBox(
-			@Arg(name = "x", description = "The x position from the top left") short x,
-			@Arg(name = "y", description = "The y position from the top left") short y,
-			@Arg(name = "width", description = "The width of the box") short width,
-			@Arg(name = "height", description = "The height of the box") short height,
+			@Arg(name = "x", description = "The x position from the top left") float x,
+			@Arg(name = "y", description = "The y position from the top left") float y,
+			@Arg(name = "width", description = "The width of the box") float width,
+			@Arg(name = "height", description = "The height of the box") float height,
 			@Arg(name = "color", description = "The color of the box") int color,
 			@Arg(name = "opacity", description = "The opacity of the box (from 0 to 1)") float alpha,
 			@Arg(name = "color2", description = "The color of the other side of the box") int color2,
@@ -62,8 +62,8 @@ public interface IDrawableContainer {
 
 	@ScriptCallable(returnTypes = ReturnType.OBJECT, description = "Add an icon of an item to the screen")
 	public Drawable addIcon(
-			@Arg(name = "x", description = "The x position from the top left") short x,
-			@Arg(name = "y", description = "The y position from the top left") short y,
+			@Arg(name = "x", description = "The x position from the top left") float x,
+			@Arg(name = "y", description = "The y position from the top left") float y,
 			@Arg(name = "id", description = "The id of the item to draw") String id,
 			@Optionals @Arg(name = "meta", description = "The meta of the item to draw") Short meta
 			);
@@ -71,10 +71,10 @@ public interface IDrawableContainer {
 	@Alias("addFluid")
 	@ScriptCallable(returnTypes = ReturnType.OBJECT, description = "Add a box textured like a liquid to the screen")
 	public Drawable addLiquid(
-			@Arg(name = "x", description = "The x position from the top left") short x,
-			@Arg(name = "y", description = "The y position from the top left") short y,
-			@Arg(name = "width", description = "The width of the liquid box") short width,
-			@Arg(name = "height", description = "The height of the liquid box") short height,
+			@Arg(name = "x", description = "The x position from the top left") float x,
+			@Arg(name = "y", description = "The y position from the top left") float y,
+			@Arg(name = "width", description = "The width of the liquid box") float width,
+			@Arg(name = "height", description = "The height of the liquid box") float height,
 			@Arg(name = "liquid", description = "The name of the fluid to render") String id
 			);
